+++ src/mlektic/logistic.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 
 import plotly.io as pio
 
-from .api.logistic import visualize_logistic
+from .api.logistic import explain_logistic_prediction, visualize_logistic
 from .services.logistic_history import fit_history_logistic
 from .visualization.logistic.binary_1d import build_binary_simple_logistic_figure
 from .visualization.logistic.binary_2d import build_binary_plane_logistic_figure
@@ -16,7 +16,6 @@ from .visualization.logistic.binary_nd import build_binary_multivar_logistic_fig
 from .visualization.logistic.multiclass_1d import build_multiclass_1d_logistic_figure
 from .visualization.logistic.multiclass_nd import build_multiclass_multivar_logistic_figure
 from .visualization.logistic.router import build_logistic_figure
-from .api.logistic import explain_logistic_prediction
 
 pio.renderers.default = "colab"